src/requirements.txt
# ========================================================================================
# Install with: pip install -r requirements.txt

# si no tienes pip, ejecutar:
# python -m pip install --upgrade pip

# Kafka client for Python
kafka-python-ng==2.2.2
